requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-groq
langchain-google-genai
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.0
numpy==1.26.0
python-dotenv==1.0.0
scipy==1.11.0
